Fix : UI 일부변경및 API로직 분리 test file 생성
