+++ src/api/estimations.ts
@@ -1,6 +1,7 @@
 import { auth } from '@/api/axiosInstance';
 import { IEstimationForm } from '@/config/types';
 
+// 견적서 GET 요청
 export const fetchEstimations = async () => {
   try {
     const response = await auth.get('/estimations/');
@@ -22,13 +23,14 @@ export const fetchExpertData = async () => {
   }
 };
 
+// 견적서 POST요청
 export const fetchEstimationsEdit = async (payload: IEstimationForm) => {
   try {
     const response = await auth.post('/estimations/request/', payload);
     console.log('response.status:', response.status);
     return response;
   } catch (error) {
-    console.error('post 요청 실패:', error);
+    console.error('POST 요청 실패:', error);
     throw error;
   }
 };
