+++ src/pages/UserEstimationPage.tsx
@@ -10,13 +10,35 @@ import { IEstimationForm, ILocation } from '@/config/types';
 import { fetchServiceLocation } from '@/api/services';
 import { useEffect, useState } from 'react';
 import WeddingLocation from '@/uiComponents/UserEstimationEdit/WeddingLocation';
-import { auth } from '@/api/axiosInstance';
+import { fetchEstimationsEdit } from '@/api/estimations';
+import { useToastStore } from '@/store/toastStore';
+import { useNavigate } from 'react-router-dom';
+import { AxiosError } from 'axios';
+
+export const findLocationKey = (
+  apiLocation: ILocation,
+  location: string | null,
+  locationDetail: string | null
+): string | undefined => {
+  if (!location) return undefined;
+  for (const [region, areas] of Object.entries(apiLocation)) {
+    if (Array.isArray(areas)) {
+      for (const area of areas) {
+        const matched = Object.entries(area).find(([key]) => key === locationDetail);
+        if (matched) return matched[1];
+      }
+    } else if (region === location) {
+      return areas; // 광역시나 특별시의 경우
+    }
+  }
+  return undefined;
+};
 
 // dirtyFields : 기본값이 수정된 필드
 // touchedFields : 사용자의 의해 수정된 필드들
 // watch : 구독 값 실시간 체크 (값에 따라 리렌더링 발생)
 // getValues : 값 반환, (리렌더링, 해당 값 추적 X)
-const UserEstimationPage = () => {
+const UEPage = () => {
   const {
     register,
     formState: { errors, dirtyFields },
@@ -35,106 +57,114 @@ const UserEstimationPage = () => {
     },
   });
 
+  const { addToasts } = useToastStore();
+  const navigate = useNavigate();
   // 지역 data api
   const [apiLocation, setApiLocation] = useState<ILocation | []>([]);
-  const locationData = async () => {
-    try {
-      const data = await fetchServiceLocation();
-      return setApiLocation(data);
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  // 렌더링시 api 호출
-  useEffect(() => {
-    locationData();
-    console.log(apiLocation);
-  }, []);
-
   // wedding location에 넘겨줄 state
   const [select, setSelect] = useState<string | null>(null);
   const [selectDetail, setSelectDetail] = useState<string | null>(null);
 
-  // POST 요청을 처리하는 함수
-  const postEstimationData = async (formData: IEstimationForm) => {
-    try {
-      // location 데이터를 매칭하는 함수
-      const findLocationKey = (location: string | null): string | undefined => {
-        if (!location) return undefined;
-        for (const [region, areas] of Object.entries(apiLocation)) {
-          if (Array.isArray(areas)) {
-            for (const area of areas) {
-              const matched = Object.entries(area).find(([k]) => k === location);
-              if (matched) return matched[1];
-            }
-          } else if (region === location) {
-            return areas; // 광역시나 특별시의 경우
-          }
-        }
-        return undefined;
-      };
-
-      // location 매칭
-      const matchedLocation = findLocationKey(select, selectDetail);
-
-      if (!matchedLocation) {
-        console.log('선택한 지역이 서버 데이터와 일치하지 않습니다.');
-        return;
+  // 렌더링시 api 호출
+  useEffect(() => {
+    const locationData = async () => {
+      try {
+        const data = await fetchServiceLocation();
+        setApiLocation(data);
+      } catch (err) {
+        console.error(err);
       }
+    };
+    locationData();
+  }, []);
 
-      // POST 요청 데이터 생성
-      const payload = {
-        service_list: Array.isArray(formData.service_list) ? formData.service_list : [formData.service_list],
-        prefer_gender: formData.prefer_gender === 'man' ? 'M' : formData.prefer_gender === 'woman' ? 'F' : 'ANY',
-        wedding_hall: formData.wedding_hall,
-        wedding_datetime: formData.wedding_datetime,
-        location: matchedLocation,
-      };
-
-      console.log('Payload:', payload);
-
-      // 서버로 POST 요청
-      const response = await auth.post('/estimations/request/', payload);
-      console.log(response, '견적 요청이 성공적으로 전송되었습니다.');
-    } catch (error) {
-      console.error('POST 요청 실패:', error);
-      console.log('견적 요청 중 오류가 발생했습니다.');
-    }
-  };
+  // POST 요청을 처리하는 함수
   // 광역시, 특별시는 선택되었을때 견적요청하기 버튼 활성화되는 조건필요
   // 캘린더, 시간도 기본값 아닌 내용 채워졌을때 버튼 활성화
   // 희망일정 선택시 금일 날짜 이후 선택
   // 제출버튼을 눌렀을때 각 칸이 채워졌는지 검증하면서 +토스트
   // 제출버튼 클릭후에는 서버응답 200일 경우 토스트 띄우고 메인
 
-  const findLocation = (location: string, locationDetail: string | null): string | void => {
-    if (location) {
-      console.log('apiLocation:', apiLocation);
-      console.log('select, selectDetail:', select, selectDetail);
-    } else if (locationDetail) {
-      return location;
-    } else if (!location) {
-      console.log('지역정보를 불러올수 없습니다.');
+  // 토스트-페이지 핸들링
+  const handleResponse = async (
+    apiCall: () => Promise<any>, // 실행할 API 호출 함수
+    successMsg: string, // 성공 시 표시할 메시지
+    errorMsg: string // 기본 에러 메시지 (서버에서 메시지를 주지 않을 경우 표시)
+  ) => {
+    try {
+      // API 호출 실행
+      const response = await apiCall();
+
+      // 상태 코드가 200인 경우
+      if (response.status === 200) {
+        addToasts({
+          type: 'success',
+          title: successMsg,
+          id: Date.now().toString(),
+        });
+        navigate('/');
+      }
+      return response;
+    } catch (error: any) {
+      const axiosError: any = error as AxiosError;
+      const responseData = axiosError.response?.data;
+
+      // 특정한 에러 메시지가 있는 경우 처리
+      if (responseData?.wedding_datetime && Array.isArray(responseData.wedding_datetime)) {
+        // wedding_datetime 에러 메시지가 존재할 경우 해당 메시지 표시
+        addToasts({
+          type: 'error',
+          title: responseData.wedding_datetime.join(', '), // 배열 메시지 병합
+          id: Date.now().toString(),
+        });
+      } else {
+        // 서버에서 detail, message를 제공시 해당 msg 사용, 없으면 기본 에러 msg 사용
+        const errorMessage = responseData?.detail || responseData?.message || errorMsg;
+
+        // 일반적인 에러 메시지 토스트로 표시
+        addToasts({
+          type: 'error',
+          title: errorMessage,
+          id: Date.now().toString(),
+        });
+      }
+
+      // 콘솔 로그로 디버깅 정보 출력 (개발용)
+      // console.error('API 호출 실패:', axiosError);
+      // console.error('응답 데이터:', responseData);
     }
   };
 
-  // 폼 제출 처리
-  const submitClick = (data: IEstimationForm) => {
-    // console.log('data:', data);
-    // // console.log(Object.entries(data));
-    // const service = watch('service_list');
-    // const gender = getValues('prefer_gender');
-    // const datetime = getValues('wedding_datetime');
-    // const hall = getValues('wedding_hall');
-    // const location = select;
-    // const locationDetail = selectDetail;
-    // console.log(service, gender, datetime, location, locationDetail, hall);
+  const submitClick = async (formData: IEstimationForm) => {
+    // 지역매칭
+    const matchedLocation = findLocationKey(apiLocation, select, selectDetail);
 
-    findLocation(select, selectDetail);
-    console.log(data);
-    // api 요청처리
-    postEstimationData(data);
+    if (!matchedLocation) {
+      console.log('선택한 지역이 서버 데이터와 일치하지 않습니다.');
+      return;
+    }
+
+    // payload 생성
+    const payload = {
+      service_list: Array.isArray(formData.service_list) ? formData.service_list : [formData.service_list],
+      prefer_gender: formData.prefer_gender === 'man' ? 'M' : formData.prefer_gender === 'woman' ? 'F' : 'ANY',
+      wedding_hall: formData.wedding_hall,
+      wedding_datetime: formData.wedding_datetime,
+      location: matchedLocation,
+    };
+
+    console.log('Payload:', payload);
+    try {
+      // handleResponse로 서버status에 따른 토스트메세지 표현
+      // POST요청(formdata 포함)
+      await handleResponse(
+        () => fetchEstimationsEdit(payload),
+        '견적 요청이 완료되었습니다.',
+        '견적 요청 처리 중 오류가 발생했습니다.'
+      );
+    } catch (error) {
+      console.error(error);
+    }
   };
 
   return (
@@ -144,8 +174,8 @@ const UserEstimationPage = () => {
         <div className='contents'>
           <MediumTitle extraClass='mainTitle' title='견적 요청서' />
           <form onSubmit={handleSubmit(submitClick)}>
-            <div className='service'>
-              <SmallTitle title='서비스 선택' />
+            <div className='serviceWrapper'>
+              <SmallTitle extraClass='service' title='서비스 선택' />
               <div className='choose'>
                 <span className='chkBox'>
                   <input
@@ -168,11 +198,11 @@ const UserEstimationPage = () => {
                   <input type='checkbox' id='singer' value='singer' {...register('service_list')} />
                   <label htmlFor='singer'>축가 가수</label>
                 </span>
-                {errors.service_list && <span className='errorMsg'>{errors.service_list.message}</span>}
               </div>
+              {errors.service_list && <div className='errorMsg'>{errors.service_list.message}</div>}
             </div>
-            <div className='gender'>
-              <SmallTitle title='전문가 선호 성별' />
+            <div className='genderWrapper'>
+              <SmallTitle extraClass='gender' title='전문가 선호 성별' />
               <div className='choose'>
                 <span className='rdoBox'>
                   <input
@@ -192,22 +222,24 @@ const UserEstimationPage = () => {
                   <label htmlFor='anyGender'>상관 없음</label>
                 </span>
               </div>
-              {errors.prefer_gender && <span className='errorMsg'>{errors.prefer_gender.message}</span>}
+              {errors.prefer_gender && <div className='errorMsg'>{errors.prefer_gender.message}</div>}
             </div>
-            <div className='dateTime'>
-              <SmallTitle title='예식 일정 및 예식 시간' />
+            <div className='dateTimeWrapper'>
+              <SmallTitle extraClass='weddingDateTime' title='예식 일정 및 예식 시간' />
               <div className='time'>
                 <input
                   className='comInput'
                   type='datetime-local'
                   id='dateTime'
-                  {...register('wedding_datetime', { required: '일정을 선택해 주세요.' })}
+                  {...register('wedding_datetime', {
+                    required: '일정을 선택해 주세요. 예정일은 오늘로부터 7일 이후여야 합니다.',
+                  })}
                 />
               </div>
-              {errors.wedding_datetime && <span className='errorMsg'>{errors.wedding_datetime.message}</span>}
+              {errors.wedding_datetime && <div className='errorMsg'>{errors.wedding_datetime.message}</div>}
             </div>
-            <div className='location'>
-              <SmallTitle title='지역 선택' />
+            <div className='locationWrapper'>
+              <SmallTitle extraClass='location' title='지역 선택' />
               <div>
                 <WeddingLocation
                   select={select}
@@ -217,24 +249,24 @@ const UserEstimationPage = () => {
                 />
               </div>
             </div>
-            <div className='hallName'>
-              <SmallTitle title='웨딩홀 이름' />
+            <div className='hallNameWrapper'>
+              <SmallTitle extraClass='hallName' title='웨딩홀 이름' />
               <input
                 className='comInput'
                 type='text'
                 placeholder='웨딩홀 이름을 작성해 주세요'
                 {...register('wedding_hall', { required: '웨딩홀 이름은 필수 항목입니다.' })}
               />
-              {errors.wedding_hall && <span className='errorMsg'>{errors.wedding_hall.message}</span>}
+              {errors.wedding_hall && <div className='errorMsg'>{errors.wedding_hall.message}</div>}
             </div>
-            <div className='method'>
-              <SmallTitle title='예식 방식' />
-              <div className='is_reception'>
+            <div className='methodWrapper'>
+              <SmallTitle extraClass='method' title='예식 방식' />
+              <div className='isReception'>
                 <label>
                   <XSmallTitle title='2부 예식을 진행하나요?' />
                 </label>
                 <div className='choose'>
-                  <span className='rdoBox'>
+                  <div className='rdoBox'>
                     <input
                       type='radio'
                       id='yes'
@@ -244,10 +276,10 @@ const UserEstimationPage = () => {
                     <label htmlFor='yes'>예</label>
                     <input type='radio' id='no' value='false' {...register('is_reception')} />
                     <label htmlFor='no'>아니오</label>
-                    {errors.is_reception && <span className='errorMsg'>{errors.is_reception.message}</span>}
-                  </span>
+                  </div>
                 </div>
               </div>
+              {errors.is_reception && <div className='errorMsg'>{errors.is_reception.message}</div>}
             </div>
             <div className='estimationBtn'>
               <MainBtn name='견적 요청하기' extraClass='extraClass' type='submit' disabled={false} />
@@ -259,4 +291,4 @@ const UserEstimationPage = () => {
   );
 };
 
-export default UserEstimationPage;
+export default UEPage;
